Change /api/blog/open/[slug] endpoint to Server Actions
