+++ src/app/blog/[slug]/edit/_actions/openInCotEditor.ts
@@ -1,0 +1,21 @@
+'use server'
+
+import { execSync } from 'child_process'
+import fs from 'fs'
+import path from 'path'
+
+export async function openInCotEditor(slug: string) {
+  if (process.env.NODE_ENV !== 'development') {
+    throw new Error('Forbidden')
+  }
+
+  const postsDirectory = path.join(process.cwd(), 'src', 'posts')
+  const fullPath = path.join(postsDirectory, `${slug}.md`)
+
+  if (fs.existsSync(fullPath)) {
+    console.log(`open ${slug}`)
+    execSync(`cot ${fullPath}`)
+  } else {
+    throw new Error('Not found')
+  }
+}
